Rework zsh config to use antibody for plugins
Switch out omz for antibody and rework which plugins are used with zsh. The
result is somewhat of a hybrid from my fish shell and zsh preferences optimized
for load time.

There's no doubt plenty of room for improvement but load times drastically
decreased. Prior to this each shell would load in between ~2 to ~3 seconds.

Current test:

$ for i in $(seq 1 10); do /usr/bin/time zsh -i -c exit; done
        0.73 real         0.22 user         0.22 sys
        0.75 real         0.32 user         0.25 sys
        0.48 real         0.23 user         0.18 sys
        0.48 real         0.22 user         0.18 sys
        0.64 real         0.28 user         0.22 sys
        0.63 real         0.29 user         0.23 sys
        0.48 real         0.22 user         0.18 sys
        0.48 real         0.22 user         0.18 sys
        0.72 real         0.32 user         0.25 sys
        0.51 real         0.24 user         0.19 sys

my goal is to get to around ~0.3

- https://getantibody.github.io/